requirements-dev.txt
numpy
matplotlib
scikit-learn
vtk
nibabel
SimpleITK
pynrrd
hcatnetwork @ git+https://github.com/AAMIASoftwares-research/HCATNetwork.git@main